BACKGROUND:
- Mesh-Ray intersection used a very brute-force method, looping over all triangles and checking them all for intersection. We were already building an octree for the mesh when adding vertices, but not using it.

DESCRIPTION:
- Uses the mesh's octree to find nearby triangles before trying to compute an intersection
- Exposes a new `Ray.NearbyPoints` method.
- Exposes a public `Elements.Search.PointOctree<T>` method as a thin wrapper around NetOctree's `PointOctree<T>`
- Enable turning the mesh's `BuildOctree` behavior on and off. With it on, mesh construction takes ~2x as long. We leave it on by default since it's still pretty fast, and we want the capabilities enabled by the octree to work "out of the box".

TESTING:
- Added a test for `Ray.NearbyPoints`
- added a new benchmark for `Ray.Intersects(mesh)`:

Before:

|                           Method |    Mean |    Error |   StdDev |
|--------------------------------- |--------:|---------:|---------:|
| 'Intersect 1000 rays with mesh.' | 2.698 s | 0.0365 s | 0.0341 s |

After:

|                           Method |     Mean |   Error |  StdDev |
|--------------------------------- |---------:|--------:|--------:|
| 'Intersect 1000 rays with mesh.' | 148.1 ms | 1.34 ms | 1.25 ms |

- added benchmarks for mesh construction at various sizes with and without the octree:

|           Method | BuildOctree | VertexCount |       Mean |     Error |    StdDev |
|----------------- |------------ |------------ |-----------:|----------:|----------:|
| 'Construct Mesh' |       False |        1000 |   1.156 ms | 0.0120 ms | 0.0112 ms |
| 'Construct Mesh' |       False |        5000 |  12.364 ms | 0.2312 ms | 0.2162 ms |
| 'Construct Mesh' |       False |       10000 |  30.029 ms | 0.6001 ms | 1.3171 ms |
| 'Construct Mesh' |       False |       30000 |  80.432 ms | 1.0678 ms | 0.8916 ms |
| 'Construct Mesh' |        True |        1000 |   3.629 ms | 0.0509 ms | 0.0476 ms |
| 'Construct Mesh' |        True |        5000 |  25.598 ms | 0.3935 ms | 0.3680 ms |
| 'Construct Mesh' |        True |       10000 |  54.057 ms | 0.6537 ms | 0.5459 ms |
| 'Construct Mesh' |        True |       30000 | 194.824 ms | 2.2398 ms | 1.9856 ms |
  
FUTURE WORK:
- This intersects method works best for meshes with more-or-less uniform triangles. If we encounter a lot of meshes with very oblong triangles, it might not be ideal. 

REQUIRED:
- [X] All changes are up to date in `CHANGELOG.md`.

<!-- Reviewable:start -->
- - -
This change is [<img src="https://reviewable.io/review_button.svg" height="34" align="absmiddle" alt="Reviewable"/>](https://reviewable.io/reviews/hypar-io/elements/849)
<!-- Reviewable:end -->