@andrewheumann 2022-08-08
So, before you review this guy, it turns out building the octree on construction can be a real performance drag for large meshes. I think we at least need a way to opt out.