@sanchez 2022-08-08
Can this be a delayed function or something that can be triggered by a function later down the chain of execution? But if we have it execute within a separate function then it would also need to be serialized too which would massively blow up Model size?