@andrewheumann 2022-08-08
@sanchez I don't think we want to serialize the Octree — I think it makes sense to build it at deserialization / mesh construction time. As it stands the `PointOctree` class from the library we're using isn't serializable in any case. I've made building the Octree optional, so you can construct a mesh quickly in one function, then read it into another function and build an octree as needed (either with `BuildOctree`= true, or by forcing it with `ForceBuildOctree()`