@sanchez 2022-08-08
Could we instead by default not build it, then the first time a query is executed that requires the octree it is built?